refactor: Move request and response structs to dtos package

+++ internal/handlers/chatroom_handler.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 
+	"github.com/teamyapchat/yapchat-server/internal/dtos"
 	"github.com/teamyapchat/yapchat-server/internal/models"
 	"github.com/teamyapchat/yapchat-server/internal/services"
 	"github.com/teamyapchat/yapchat-server/internal/utils"
@@ -29,27 +30,13 @@ func NewChatRoomHandler(
 	}
 }
 
-// ChatRoomResponse defines the response structure for chat room related API calls
-type ChatRoomResponse struct {
-	ID           uint           `json:"id"`
-	Name         string         `json:"name"`
-	Type         string         `json:"type"`
-	Participants []UserResponse `json:"participants"`
-}
-
-type MessageResponse struct {
-	Content   string `json:"content"`
-	SenderID  string `json:"sender_id"`
-	Timestamp string `json:"timestamp"`
-}
-
 // CreateHandler godoc
 // @Summary      Create a new chat room
 // @Description  Create a new chat room
 // @Tags         chatrooms
 // @Accept       json
 // @Produce      json
-// @Param        request body models.ChatRoomRequest true "Chat room info"
+// @Param        request body dtos.ChatRoomRequest true "Chat room info"
 // @Success      201 {object} utils.SuccessResponse
 // @Failure      400 {object} utils.ErrorResponse
 // @Failure      401 {object} utils.ErrorResponse
@@ -62,7 +49,7 @@ func (h *ChatRoomHandler) CreateHandler(c *gin.Context) {
 		return
 	}
 
-	var chatroomRequest models.ChatRoomRequest
+	var chatroomRequest dtos.ChatRoomRequest
 	if err := c.ShouldBindJSON(&chatroomRequest); err != nil {
 		c.JSON(http.StatusBadRequest, utils.NewErrorResponse("Invalid request body"))
 		return
@@ -75,7 +62,12 @@ func (h *ChatRoomHandler) CreateHandler(c *gin.Context) {
 		return
 	}
 
-	c.JSON(http.StatusCreated, utils.NewSuccessResponse("Chat room created successfully"))
+	c.JSON(http.StatusCreated,
+		utils.SuccessResponse{
+			Success: true,
+			Message: "Chat room created successfully",
+		},
+	)
 }
 
 // GetByIDHandler godoc
@@ -84,7 +76,7 @@ func (h *ChatRoomHandler) CreateHandler(c *gin.Context) {
 // @Tags         chatrooms
 // @Produce      json
 // @Param        id path integer true "Chat room ID"
-// @Success      200 {object} utils.SuccessResponse{data=ChatRoomResponse}
+// @Success      200 {object} utils.SuccessResponse{data=dtos.ChatRoomResponse}
 // @Failure      400 {object} utils.ErrorResponse
 // @Failure      401 {object} utils.ErrorResponse
 // @Failure      404 {object} utils.ErrorResponse
@@ -120,11 +112,12 @@ func (h *ChatRoomHandler) GetByIDHandler(c *gin.Context) {
 		return
 	}
 
-	response := ChatRoomResponse{
+	response := dtos.ChatRoomResponse{
 		ID:           chatroom.ID,
 		Name:         chatroom.Name,
 		Type:         string(chatroom.Type),
 		Participants: getParticipants(chatroom.Participants),
+		ImageURL:     chatroom.ImageURL,
 	}
 
 	c.JSON(http.StatusOK, utils.NewSuccessResponse(response))
@@ -137,7 +130,7 @@ func (h *ChatRoomHandler) GetByIDHandler(c *gin.Context) {
 // @Produce      json
 // @Param        id path integer true "Chat room ID"
 // @Param        count query integer false "Number of messages to return (default 25)"
-// @Success      200 {object} utils.SuccessResponse{data=[]MessageResponse}
+// @Success      200 {object} utils.SuccessResponse{data=[]dtos.MessageResponse}
 // @Failure      400 {object} utils.ErrorResponse
 // @Failure      401 {object} utils.ErrorResponse
 // @Failure      403 {object} utils.ErrorResponse
@@ -185,9 +178,9 @@ func (h *ChatRoomHandler) GetMessagesByRoomIDHandler(c *gin.Context) {
 		return
 	}
 
-	messageList := make([]MessageResponse, 0, len(messages))
+	messageList := make([]dtos.MessageResponse, 0, len(messages))
 	for _, message := range messages {
-		messageList = append(messageList, MessageResponse{
+		messageList = append(messageList, dtos.MessageResponse{
 			Content:   message.Content,
 			SenderID:  message.SenderID,
 			Timestamp: message.Timestamp.Format(time.RFC3339),
@@ -202,7 +195,7 @@ func (h *ChatRoomHandler) GetMessagesByRoomIDHandler(c *gin.Context) {
 // @Description  Get a list of all chat rooms that the user is in
 // @Tags         chatrooms
 // @Produce      json
-// @Success      200 {object} utils.SuccessResponse{data=[]ChatRoomResponse}
+// @Success      200 {object} utils.SuccessResponse{data=[]dtos.ChatRoomResponse}
 // @Failure      401 {object} utils.ErrorResponse
 // @Failure      500 {object} utils.ErrorResponse
 // @Router       /v1/chatrooms [get]
@@ -219,15 +212,16 @@ func (h *ChatRoomHandler) ListChatroomsHandler(c *gin.Context) {
 		return
 	}
 
-	var responses []ChatRoomResponse
+	var responses []dtos.ChatRoomResponse
 	for _, chatroom := range chatrooms {
 		participants := getParticipants(chatroom.Participants)
 
-		responses = append(responses, ChatRoomResponse{
+		responses = append(responses, dtos.ChatRoomResponse{
 			ID:           chatroom.ID,
 			Name:         chatroom.Name,
 			Type:         string(chatroom.Type),
 			Participants: participants,
+			ImageURL:     chatroom.ImageURL,
 		})
 	}
 
@@ -299,6 +293,7 @@ func (h *ChatRoomHandler) GetInviteCodeHandler(c *gin.Context) {
 // @Failure      400 {object} utils.ErrorResponse
 // @Failure      401 {object} utils.ErrorResponse
 // @Failure      404 {object} utils.ErrorResponse
+// @Failure      409 {object} utils.ErrorResponse
 // @Failure      500 {object} utils.ErrorResponse
 // @Router       /v1/chatrooms/{id}/join [post]
 func (h *ChatRoomHandler) JoinChatroomHandler(c *gin.Context) {
@@ -333,6 +328,11 @@ func (h *ChatRoomHandler) JoinChatroomHandler(c *gin.Context) {
 		return
 	}
 
+	if slices.Contains(getParticipantIDs(chatroom.Participants), userID.(string)) {
+		c.JSON(http.StatusConflict, utils.NewErrorResponse("User already in chat room"))
+		return
+	}
+
 	err = h.chatroomService.AddParticipant(chatroomID, userID.(string))
 	if err != nil {
 		if errors.Is(err, services.ErrChatRoomNotFound) {
@@ -357,10 +357,11 @@ func (h *ChatRoomHandler) JoinChatroomHandler(c *gin.Context) {
 // @Tags         chatrooms
 // @Produce      json
 // @Param        id path integer true "Chat room ID"
-// @Success      200 {object} utils.SuccessResponse
+// @Success      204
 // @Failure      400 {object} utils.ErrorResponse
 // @Failure      401 {object} utils.ErrorResponse
 // @Failure      404 {object} utils.ErrorResponse
+// @Failure      409 {object} utils.ErrorResponse
 // @Failure      500 {object} utils.ErrorResponse
 // @Router       /v1/chatrooms/{id}/leave [post]
 func (h *ChatRoomHandler) LeaveChatroomHandler(c *gin.Context) {
@@ -379,6 +380,20 @@ func (h *ChatRoomHandler) LeaveChatroomHandler(c *gin.Context) {
 	chatroomID := uint(idUint64)
 
 	// TODO: Only remove user that exists in chat room
+	chatroom, err := h.chatroomService.GetByID(chatroomID)
+	if err != nil {
+		if errors.Is(err, services.ErrChatRoomNotFound) {
+			c.JSON(http.StatusNotFound, utils.NewErrorResponse("Chat room not found"))
+		} else {
+			c.JSON(http.StatusInternalServerError, utils.NewErrorResponse("Failed to leave chat room"))
+		}
+		return
+	}
+
+	if !slices.Contains(getParticipantIDs(chatroom.Participants), userID.(string)) {
+		c.JSON(http.StatusConflict, utils.NewErrorResponse("User not in chat room"))
+		return
+	}
 
 	err = h.chatroomService.RemoveParticipant(chatroomID, userID.(string))
 	if err != nil {
@@ -390,13 +405,13 @@ func (h *ChatRoomHandler) LeaveChatroomHandler(c *gin.Context) {
 		return
 	}
 
-	c.JSON(http.StatusOK, utils.NewSuccessResponse("Successfully left chat room"))
+	c.Status(http.StatusNoContent)
 }
 
-func getParticipants(participants []*models.User) []UserResponse {
-	users := make([]UserResponse, 0, len(participants))
+func getParticipants(participants []*models.User) []dtos.UserResponse {
+	users := make([]dtos.UserResponse, 0, len(participants))
 	for _, p := range participants {
-		users = append(users, UserResponse{
+		users = append(users, dtos.UserResponse{
 			ID:       p.ID,
 			Username: p.Username,
 			ImageURL: p.ImageURL,
